style : BestProductVO cnt 타입 수정

+++ product-service/src/main/java/com/lotfresh/productservice/domain/product/vo/BestProductVO.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class BestProductVO {
   private Long ProductId;
-  private Integer cnt;
+  private Long cnt;
 }
